Many Updates
Added idle screen image. Updated deploy batch file. Added
initializations for door motor controller which will close the door and
carousel motor controller which will locate which figure out and set
what disc it's on. Also added functionality to spin carousel to disc
number. The door is opened if it's not already when the disc is done
being brought to the font.